Write directly to file.
