+++ cli/standalone/binary.rs
@@ -5,6 +5,7 @@ use std::io::Seek;
 use std::io::SeekFrom;
 use std::io::Write;
 use std::path::Path;
+use std::sync::Arc;
 
 use deno_ast::ModuleSpecifier;
 use deno_core::anyhow::Context;
@@ -18,6 +19,15 @@ use deno_runtime::permissions::PermissionsOptions;
 use log::Level;
 use serde::Deserialize;
 use serde::Serialize;
+
+use crate::args::CaData;
+use crate::args::CliOptions;
+use crate::args::CompileFlags;
+use crate::cache::DenoDir;
+use crate::file_fetcher::FileFetcher;
+use crate::http_util::HttpClient;
+use crate::util::progress_bar::ProgressBar;
+use crate::util::progress_bar::ProgressBarStyle;
 
 const MAGIC_TRAILER: &[u8; 8] = b"d3n0l4nd";
 
@@ -37,13 +47,14 @@ pub struct Metadata {
   pub entrypoint: ModuleSpecifier,
 }
 
-pub fn get_binary_bytes(
-  mut original_bin: Vec<u8>,
+pub fn write_binary_bytes(
+  writer: &mut impl Write,
+  original_bin: Vec<u8>,
   metadata: &Metadata,
   eszip: eszip::EszipV2,
-) -> Result<Vec<u8>, AnyError> {
-  let mut metadata = serde_json::to_string(metadata)?.as_bytes().to_vec();
-  let mut eszip_archive = eszip.into_bytes();
+) -> Result<(), AnyError> {
+  let metadata = serde_json::to_string(metadata)?.as_bytes().to_vec();
+  let eszip_archive = eszip.into_bytes();
 
   let eszip_pos = original_bin.len();
   let metadata_pos = eszip_pos + eszip_archive.len();
@@ -51,18 +62,15 @@ pub fn get_binary_bytes(
   trailer.write_all(&eszip_pos.to_be_bytes())?;
   trailer.write_all(&metadata_pos.to_be_bytes())?;
 
-  let mut final_bin = Vec::with_capacity(
-    original_bin.len() + eszip_archive.len() + metadata.len() + trailer.len(),
-  );
-  final_bin.append(&mut original_bin);
-  final_bin.append(&mut eszip_archive);
-  final_bin.append(&mut metadata);
-  final_bin.append(&mut trailer);
+  writer.write_all(&original_bin)?;
+  writer.write_all(&eszip_archive)?;
+  writer.write_all(&metadata)?;
+  writer.write_all(&trailer)?;
 
-  Ok(final_bin)
+  Ok(())
 }
 
-pub fn is_compiled_binary(exe_path: &Path) -> bool {
+pub fn is_standalone_binary(exe_path: &Path) -> bool {
   let Ok(mut output_file) = std::fs::File::open(exe_path) else {
     return false;
   };
@@ -140,4 +148,160 @@ fn u64_from_bytes(arr: &[u8]) -> Result<u64, AnyError> {
     .try_into()
     .context("Failed to convert the buffer into a fixed-size array")?;
   Ok(u64::from_be_bytes(*fixed_arr))
+}
+
+pub struct DenoCompileBinaryWriter {
+  file_fetcher: Arc<FileFetcher>,
+  client: HttpClient,
+  deno_dir: DenoDir,
+}
+
+impl DenoCompileBinaryWriter {
+  pub fn new(
+    file_fetcher: Arc<FileFetcher>,
+    client: HttpClient,
+    deno_dir: DenoDir,
+  ) -> Self {
+    Self {
+      file_fetcher,
+      client,
+      deno_dir,
+    }
+  }
+
+  pub async fn write_bin(
+    &self,
+    writer: &mut impl Write,
+    eszip: eszip::EszipV2,
+    module_specifier: &ModuleSpecifier,
+    compile_flags: &CompileFlags,
+    cli_options: &CliOptions,
+  ) -> Result<(), AnyError> {
+    // Select base binary based on target
+    let original_binary =
+      self.get_base_binary(compile_flags.target.clone()).await?;
+
+    self
+      .write_standalone_binary(
+        writer,
+        original_binary,
+        eszip,
+        module_specifier,
+        cli_options,
+        compile_flags,
+      )
+      .await
+  }
+
+  async fn get_base_binary(
+    &self,
+    target: Option<String>,
+  ) -> Result<Vec<u8>, AnyError> {
+    if target.is_none() {
+      let path = std::env::current_exe()?;
+      return Ok(std::fs::read(path)?);
+    }
+
+    let target = target.unwrap_or_else(|| env!("TARGET").to_string());
+    let binary_name = format!("deno-{target}.zip");
+
+    let binary_path_suffix = if crate::version::is_canary() {
+      format!("canary/{}/{}", crate::version::GIT_COMMIT_HASH, binary_name)
+    } else {
+      format!("release/v{}/{}", env!("CARGO_PKG_VERSION"), binary_name)
+    };
+
+    let download_directory = self.deno_dir.dl_folder_path();
+    let binary_path = download_directory.join(&binary_path_suffix);
+
+    if !binary_path.exists() {
+      self
+        .download_base_binary(&download_directory, &binary_path_suffix)
+        .await?;
+    }
+
+    let archive_data = std::fs::read(binary_path)?;
+    let temp_dir = tempfile::TempDir::new()?;
+    let base_binary_path = crate::tools::upgrade::unpack_into_dir(
+      archive_data,
+      target.contains("windows"),
+      &temp_dir,
+    )?;
+    let base_binary = std::fs::read(base_binary_path)?;
+    drop(temp_dir); // delete the temp dir
+    Ok(base_binary)
+  }
+
+  async fn download_base_binary(
+    &self,
+    output_directory: &Path,
+    binary_path_suffix: &str,
+  ) -> Result<(), AnyError> {
+    let download_url = format!("https://dl.deno.land/{binary_path_suffix}");
+    let maybe_bytes = {
+      let progress_bars = ProgressBar::new(ProgressBarStyle::DownloadBars);
+      let progress = progress_bars.update(&download_url);
+
+      self
+        .client
+        .download_with_progress(download_url, &progress)
+        .await?
+    };
+    let bytes = match maybe_bytes {
+      Some(bytes) => bytes,
+      None => {
+        log::info!("Download could not be found, aborting");
+        std::process::exit(1)
+      }
+    };
+
+    std::fs::create_dir_all(output_directory)?;
+    let output_path = output_directory.join(binary_path_suffix);
+    std::fs::create_dir_all(output_path.parent().unwrap())?;
+    tokio::fs::write(output_path, bytes).await?;
+    Ok(())
+  }
+
+  /// This functions creates a standalone deno binary by appending a bundle
+  /// and magic trailer to the currently executing binary.
+  async fn write_standalone_binary(
+    &self,
+    writer: &mut impl Write,
+    original_bin: Vec<u8>,
+    eszip: eszip::EszipV2,
+    entrypoint: &ModuleSpecifier,
+    cli_options: &CliOptions,
+    compile_flags: &CompileFlags,
+  ) -> Result<(), AnyError> {
+    let ca_data = match cli_options.ca_data() {
+      Some(CaData::File(ca_file)) => Some(
+        std::fs::read(ca_file)
+          .with_context(|| format!("Reading: {ca_file}"))?,
+      ),
+      Some(CaData::Bytes(bytes)) => Some(bytes.clone()),
+      None => None,
+    };
+    let maybe_import_map = cli_options
+      .resolve_import_map(&self.file_fetcher)
+      .await?
+      .map(|import_map| (import_map.base_url().clone(), import_map.to_json()));
+    let metadata = Metadata {
+      argv: compile_flags.args.clone(),
+      unstable: cli_options.unstable(),
+      seed: cli_options.seed(),
+      location: cli_options.location_flag().clone(),
+      permissions: cli_options.permissions_options(),
+      v8_flags: cli_options.v8_flags().clone(),
+      unsafely_ignore_certificate_errors: cli_options
+        .unsafely_ignore_certificate_errors()
+        .clone(),
+      log_level: cli_options.log_level(),
+      ca_stores: cli_options.ca_stores().clone(),
+      ca_data,
+      entrypoint: entrypoint.clone(),
+      maybe_import_map,
+    };
+
+    write_binary_bytes(writer, original_bin, &metadata, eszip)
+  }
 }
